SUP 2959 push-pull command
