+++ app/lib/bk/compat/parsers/circleci.rb
@@ -6,7 +6,6 @@ require_relative '../models/steps/group'
 require_relative 'circleci/executors'
 require_relative 'circleci/jobs'
 require_relative 'circleci/orbs'
-require_relative 'circleci/orbs/docker'
 require_relative 'circleci/steps'
 require_relative 'circleci/workflows'
